+++ All In/ContentView.swift
@@ -18,6 +18,7 @@ struct ContentView: View {
             Group {
                 if tabNavigationManager.selectedTab == .home {
                     MainProfileView(user: User.dummyData[0])
+                    HomeView(user: User.dummyData[0])
                         .transition(transitionModifier)
                 } else if tabNavigationManager.selectedTab == .market {
                     MarketplaceView()
@@ -37,6 +38,7 @@ struct ContentView: View {
         .ignoresSafeArea()
         .background(Constants.Colors.background)
     }
+
 }
 
 #Preview {
