+++ All In/Features/Home/PlayerSeeAllScreen.swift
@@ -16,11 +16,14 @@ struct PlayerSeeAllScreen: View {
     @EnvironmentObject var tabNavigationManager: TabNavigationManager
 
     // States for continuous scrolling animation
-    @State private var rightScrollOffset: CGFloat = 0
-    @State private var leftScrollOffset: CGFloat = 0
+    @State private var rowOffsets: [CGFloat] = [0, 0, 0, 0]
 
-    // Timer for automatic scrolling
-    @State private var scrollTimer: Timer?
+    // Track user interaction with gesture
+    @State private var userScrolling: [Bool] = [false, false, false, false]
+    @State private var userScrollOffset: [CGFloat] = [0, 0, 0, 0]
+
+    // Display Link for smoother animation
+    @State private var displayLink: Timer?
 
     init(sport: Sport = Sport.all.first(where: { $0.name == "Basketball" }) ?? Sport.all[0]) {
         self._selectedSport = State(initialValue: sport)
@@ -32,23 +35,28 @@ struct PlayerSeeAllScreen: View {
         let totalWidth = CGFloat(Player.dummyData.count) * playerWidth
 
         // Initialize offsets
-        rightScrollOffset = 0
-        leftScrollOffset = -totalWidth
+        for i in 0..<rowOffsets.count {
+            rowOffsets[i] = i % 2 == 0 ? 0 : -totalWidth
+        }
 
-        // Create a timer for continuous scrolling
-        scrollTimer = Timer.scheduledTimer(withTimeInterval: 0.05, repeats: true) { _ in
-            // Update scroll offsets
-            withAnimation {
-                // Right-moving row (first row)
-                rightScrollOffset -= 1
-                if rightScrollOffset < -totalWidth {
-                    rightScrollOffset = 0
-                }
-
-                // Left-moving row (second row)
-                leftScrollOffset += 1
-                if leftScrollOffset > 0 {
-                    leftScrollOffset = -totalWidth
+        // Create a display link for smooth animation
+        displayLink = Timer.scheduledTimer(withTimeInterval: 1/60, repeats: true) { _ in
+            // Update scroll offsets for rows not being interacted with
+            for i in 0..<rowOffsets.count {
+                if !userScrolling[i] {
+                    if i % 2 == 0 {
+                        // Right-moving row
+                        rowOffsets[i] -= 0.5
+                        if rowOffsets[i] < -totalWidth {
+                            rowOffsets[i] = 0
+                        }
+                    } else {
+                        // Left-moving row
+                        rowOffsets[i] += 0.5
+                        if rowOffsets[i] > 0 {
+                            rowOffsets[i] = -totalWidth
+                        }
+                    }
                 }
             }
         }
@@ -73,13 +81,12 @@ struct PlayerSeeAllScreen: View {
                 }
                 .padding(24)
 
-                ScrollView {
-                    VStack(spacing: 20) {
-                        ForEach(0..<4) { rowIndex in
-                            scrollingPlayerRow(isRightDirection: rowIndex % 2 == 0)
-                        }
+                VStack(spacing: 20) {
+                    ForEach(0..<4) { rowIndex in
+                        scrollablePlayerRow(rowIndex: rowIndex)
                     }
                 }
+                .padding(.bottom, 24)
             }
             .background(Constants.Colors.background)
             .scrollIndicators(.hidden)
@@ -91,7 +98,7 @@ struct PlayerSeeAllScreen: View {
             }
             .onDisappear {
                 tabNavigationManager.hideTabBar = false
-                scrollTimer?.invalidate()
+                displayLink?.invalidate()
             }
         }
     }
@@ -110,13 +117,16 @@ struct PlayerSeeAllScreen: View {
         }
     }
 
-    private func scrollingPlayerRow(isRightDirection: Bool) -> some View {
-        // Create a container for continuous scrolling effect
-        GeometryReader { geometry in
+    private func scrollablePlayerRow(rowIndex: Int) -> some View {
+        let isRightDirection = rowIndex % 2 == 0
+
+        return GeometryReader { geometry in
             let rowWidth = geometry.size.width
+            let playerWidth: CGFloat = 100 + 16 // Card width + spacing
+            let totalWidth = CGFloat(Player.dummyData.count) * playerWidth
 
             HStack(spacing: 16) {
-                // Duplicate the players
+                // Duplicate the players for continuous effect
                 ForEach(0..<2) { _ in
                     HStack(spacing: 16) {
                         ForEach(Player.dummyData) { player in
@@ -126,9 +136,41 @@ struct PlayerSeeAllScreen: View {
                     }
                 }
             }
-            // Apply the scrolling offset based on direction
-            .offset(x: isRightDirection ? rightScrollOffset : leftScrollOffset)
-            .frame(height: 140)
+            .offset(x: rowOffsets[rowIndex] + userScrollOffset[rowIndex])
+            .gesture(
+                DragGesture()
+                    .onChanged { value in
+                        userScrolling[rowIndex] = true
+                        userScrollOffset[rowIndex] = value.translation.width
+                    }
+                    .onEnded { _ in
+                        // Update the base offset with the final user gesture
+                        rowOffsets[rowIndex] += userScrollOffset[rowIndex]
+
+                        // Reset user offset
+                        userScrollOffset[rowIndex] = 0
+
+                        // Handle edge cases
+                        if isRightDirection {
+                            if rowOffsets[rowIndex] > 0 {
+                                rowOffsets[rowIndex] = 0
+                            } else if rowOffsets[rowIndex] < -totalWidth {
+                                rowOffsets[rowIndex] = -totalWidth
+                            }
+                        } else {
+                            if rowOffsets[rowIndex] < -totalWidth {
+                                rowOffsets[rowIndex] = -totalWidth
+                            } else if rowOffsets[rowIndex] > 0 {
+                                rowOffsets[rowIndex] = 0
+                            }
+                        }
+
+                        // Resume animation after a short delay
+                        DispatchQueue.main.asyncAfter(deadline: .now() + 0.5) {
+                            userScrolling[rowIndex] = false
+                        }
+                    }
+            )
         }
         .frame(height: 140)
         .clipped()
