+++ All In/Features/Profile/BetTrackerView.swift
@@ -232,7 +232,7 @@ struct BetTrackerView: View {
                     "F",
                     "S",
                     "SU"
-                ] : Array(1...12).map { "\($0)" }) { value in
+                ] : Array(1...4).map { "\($0)" }) { value in
                     AxisValueLabel {
                         if let day = value.as(String.self) {
                             Text(day)
